Styling, All html pages, Minor Additions, Can't Connect DB
Brian's code for connecting to his database isn't working.
I kept the file within private in a sub folder and put the orginial
config file in instead.

I made changes to the overall layout of the site. Still not perfect.

Added all the necessary pages
and modified links
-Inventory(add items)
-Suppliers(select suppliers)
-Stats

I left bar chart as is.

Created a button in inventory for adding an item.

Created classes to easily format page
- .row-content, and several deviations of row-content